+++ Empire/ResourcePool.h
@@ -28,13 +28,13 @@ public:
     const std::vector<int>&         ObjectIDs() const;                      ///< returns UniverseObject IDs in this ResourcePool
     int                             StockpileObjectID() const;              ///< returns ID of object that contains this ResourcePool's stockpile
 
-    double                          Stockpile() const;                      ///< returns current stockpiled amount of resource
-    double                          Production() const;                     ///< returns amount of resource being produced by all ResourceCenter
-    double                          GroupProduction(int object_id) const;   ///< returns amount of resource being produced by resource sharing group that contains the object with id \a object_id
+    float                           Stockpile() const;                      ///< returns current stockpiled amount of resource
+    float                           Production() const;                     ///< returns amount of resource being produced by all ResourceCenter
+    float                           GroupProduction(int object_id) const;   ///< returns amount of resource being produced by resource sharing group that contains the object with id \a object_id
 
-    double                          TotalAvailable() const;                 ///< returns amount of resource immediately available = production + stockpile from all ResourceCenters, ignoring limitations of connections between centers
-    std::map<std::set<int>, double> Available() const;                      ///< returns the sets of groups of objects that can share resources, and the amount of this pool's resource that each group has available
-    double                          GroupAvailable(int object_id) const;    ///< returns amount of resource available in resource sharing group that contains the object with id \a object_id
+    float                           TotalAvailable() const;                 ///< returns amount of resource immediately available = production + stockpile from all ResourceCenters, ignoring limitations of connections between centers
+    std::map<std::set<int>, float>  Available() const;                      ///< returns the sets of groups of objects that can share resources, and the amount of this pool's resource that each group has available
+    float                           GroupAvailable(int object_id) const;    ///< returns amount of resource available in resource sharing group that contains the object with id \a object_id
 
     std::string                     Dump() const;
     //@}
@@ -53,7 +53,7 @@ public:
      * resources are saved from turn to turn. */
     void        SetStockpileObject(int stockpile_object_id);
 
-    void        SetStockpile(double d);     ///< sets current sockpiled amount of resource
+    void        SetStockpile(float d);      ///< sets current sockpiled amount of resource
 
     void        Update();                   ///< recalculates total resource production
     //@}
@@ -63,9 +63,9 @@ private:
 
     std::vector<int>                        m_object_ids;                                   ///< IDs of objects to consider in this pool
     std::set<std::set<int> >                m_connected_system_groups;                      ///< sets of systems between and in which objects can share this pool's resource
-    std::map<std::set<int>, double>         m_connected_object_groups_resource_production;  ///< cached map from connected group of objects that can share resources, to how much resource is produced by ResourceCenters in the group.  regenerated during update from other state information.
+    std::map<std::set<int>, float>          m_connected_object_groups_resource_production;  ///< cached map from connected group of objects that can share resources, to how much resource is produced by ResourceCenters in the group.  regenerated during update from other state information.
     int                                     m_stockpile_object_id;                          ///< object id where stockpile for this pool is located
-    double                                  m_stockpile;                                    ///< current stockpiled amount of resource
+    float                                   m_stockpile;                                    ///< current stockpiled amount of resource
     ResourceType                            m_type;                                         ///< what kind of resource does this pool hold?
 
     friend class boost::serialization::access;
@@ -83,8 +83,8 @@ public:
     /** \name Accessors */ //@{
     const std::vector<int>& PopCenterIDs() const {return m_pop_center_ids;}  ///< returns the PopCenter vector
 
-    double                  Population() const;     ///< returns current total population
-    double                  Growth() const;         ///< returns predicted growth for next turn
+    float   Population() const;     ///< returns current total population
+    float   Growth() const;         ///< returns predicted growth for next turn
     //@}
 
     /** \name Mutators */ //@{
@@ -97,8 +97,8 @@ public:
 
 private:
     std::vector<int>    m_pop_center_ids;   ///< UniverseObject ids of PopCenters that contribute to the pool
-    double              m_population;       ///< total population of all PopCenters in pool
-    double              m_growth;           ///< total predicted population growth for next turn for all PopCenters in pool
+    float               m_population;       ///< total population of all PopCenters in pool
+    float               m_growth;           ///< total predicted population growth for next turn for all PopCenters in pool
 
     friend class boost::serialization::access;
     template <class Archive>
